requirements.txt
Django==1.7.1
dj-database-url==0.3.0
dj-static==0.0.6
django-appconf==0.6
django-compressor==1.4
django-debug-toolbar==1.3.2
djangorestframework==3.0.0
drf-nested-routers==0.9.0
gunicorn==19.1.1
six==1.8.0
sqlparse==0.1.15
static3==0.5.1
